refactor: changed some folders location
