+++ src/@shared/result.ts
