+++ src/application/services/auth-service.interface.ts
@@ -1,5 +1,5 @@
-import { User } from '../../../domain/entities/user';
-import { Payload } from '../models/payload';
+import { Payload } from '../../@shared/payload';
+import { User } from '../../domain/entities/user';
 
 export interface IAuthService {
     generateToken(user: User): Promise<string>;
