+++ src/application/useCases/addOperation/add-operation.interface.ts
@@ -1,5 +1,5 @@
+import { Result } from '../../../@shared/result';
 import { Operation } from '../../../domain/entities/operation';
-import { Result } from '../../../infra/models/result';
 import { AddOperationDto } from './add-operation.dto';
 
 export interface IAddOperation {
