+++ src/application/useCases/signup/signup.interface.ts
@@ -1,5 +1,5 @@
+import { Result } from '../../../@shared/result';
 import { User } from '../../../domain/entities/user';
-import { Result } from '../../../infra/models/result';
 import { SignupDto } from './signup.dto';
 
 export interface ISignup {
