--- src/info/models/info.ts
@@ -1,4 +1,0 @@
-export interface Info {
-    name?: string;
-    version?: string;
-}
