--- src/info/services/info-service.interface.ts
@@ -1,5 +1,0 @@
-import { Info } from '../models/info';
-
-export interface IInfoService {
-    getInfo(): Info;
-}
