--- src/info/types/info.types.ts
@@ -1,5 +1,0 @@
-const InfoTypes = {
-  InfoService: Symbol.for('InfoService')
-};
-
-export default InfoTypes;
