+++ src/infra/factories/info-routes.factory.ts
@@ -1,6 +1,6 @@
-import { InfoService } from '../../info/services/info.service';
 import { InfoController } from '../controllers/info.controller';
 import { InfoRoutes } from '../routes/info.routes';
+import { InfoService } from '../services/info.service';
 
 export const makeInfoRoutes = (): InfoRoutes => {
   const infoService = new InfoService();
