+++ src/infra/services/encript.service.ts
@@ -1,5 +1,5 @@
 import md5 from 'md5';
-import { IEncriptService } from '../../application/@shared/services/encript-service.interface';
+import { IEncriptService } from '../../application/services/encript-service.interface';
 import config from '../../config';
 
 export class EncriptService implements IEncriptService {
